Code of Digital Storage Oscilloscope
The project I had conduct during my master degree is to build and test
a 20MHz USB Digital Storage Oscilloscope and achieve visualization via
PC, which was a semester long project in the subject “Electronic System
Design”. The project commenced from July, 2016 to November, 2016 and
was carried out in the electrical and electronic laboratory of the
University of Melbourne. This was an individual project that apply the
electrical knowledge, engineering ability and professional attributes
into this project.